Refactor agent session surfaces
Split the React agent session list provider and the Rust agent command/service layers into focused modules while keeping public entrypoints stable.
